+++ js/data.js
@@ -1,0 +1,43 @@
+
+import { getRandomNum, getRandomFloatingPoint } from './util.js';
+
+const photosArr = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+const featuresArr = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const checkinArr = ['12:00', '13:00', '14:00'];
+const types = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const titles = ['Сдам уютную квартиру на длительный срок от собственника.', 'Звоните квартирка ждет именно Вас.', 'О квартире.', 'Об апартаментах.', 'На длительный срок впервые сдается уютная, солнечная квартира с видом на Неву.', 'Сдается просторная однушка в отличном состоянии', 'Сдается квартира с минимум мебели!', 'Большая квартира с отличным ремонтом!', 'Сдаётся отличная квартира в Ст. Петергофе.', 'Предлагается в аренду просторная квартира.', 'Сдам в Санкт-Петербурге новую уютную квартиру.', 'От собственника.', 'БЕЗ КОМИССИИ сдам современную квартиру.', 'Теплая и солнечная квартира в тихом и обустроенном спальном районе.', 'Сдается уютная видовая квартирка с качественным ремонтом в новом элитном ЖК Море.', 'Сдается светлая квартира с большой лоджией.'];
+const descriptions = ['Имеются все для комфортного проживания: кровать, гардеробная, комод с зеркалами, два раскладных дивана и другая мебель. Также есть полный кухонный гарнитур, холодильник, вся бытовая техника.', 'Просторная ванная комната и туалет совмещённые, раковина современная двойная ИКЕА, имеется биде. Вся сантехника новая.', 'Квартира сдается строго без животных.', 'В квартире курить запрещено.', 'Во дворе достаточно мест для парковки, также имеется подземный платный паркинг и открытая платная охраняемая стоянка. ', 'В шаговой доступности парк Сосновка, фитнес клуб, бассейн, магазин Окей, рестораны, кинотеатр и городская зона отдыха с озерами.', 'Квартплата по квитанции, коммунальные по счетчикам оплачивает квартиросъемщик.', 'Квартира-студия от собственника. ЖК Полис на комендантском. Хороший ремонт, кухня, стиральная машина, лоджия, большой шкаф-купе, в комнате теплые полы.', 'Прекрасное расположение комплекса в 7 минутах от метро Площадь Мужества.', 'Окна выходят на южную сторону. Без комиссии, напрямую от собственника. КУ оплата по счетчикам. Агентства прошу не беспокоить.', 'УБЕДИТЕЛЬНАЯ ПРОСЬБА - НЕ ЗВОНИТЬ ( 99% что я не смогу ответить), ПОЖАЛУЙСТА ПИШИТЕ СООБЩЕНИЯ. ', 'Я собственник, сдаю свою квартиру на полгода (возможно продление до года). Прошу вас внимательно ознакомиться с описанием и обязательно посмотреть видео.', 'В шаговой доступности садик, школа, магазины. Автобусная остановка рядом с домом, до ст.м. «Звездная», «Купчино» 20 мин.,9 км до кольцевой автодороги,1,5 км. до ж/д «Царское село».', 'Просмотр квартиры будет проводиться в виде собеседования. Я назначу день (точно будет выходной), и в часовом интервале буду показывать квартиру разным людям. В конце дня определюсь и выберу кандидатов для проживания. Опять же без каких-либо обид. Мои требования не завышены. Мне важно найти ответственных людей/ответственного человека. Ради 1 человека я мотаться не стану, прошу простить и принять. К сожалению придется подстраиваться под меня.', 'Чего в квартире нет: посуды (есть штатный набор вилок и ложек, пару чашек и тарелок, кастрюля), постельного белья, полотенец, гладильной доски. Вешалок всего 10 штук. Сушилка для белья есть.', 'Отличная транспортная развязка, с выездом на ЗСД и КАД. До м. Беговая 20 минут пешком. До м. Старая деревня 10 минут транспортом. Рядом парк 300-летия Санкт-Петербурга.'];
+
+const createAdvertArchive  = function(amount) {
+  const arr = [];
+  for (let i = 1; i <= amount; i++) {
+    const advert = {
+      author :  {avatar: (i < 10) ? `img/avatars/user0${ i }.png` : `img/avatars/user${ i }.png`},
+
+      offer : {
+        title: titles[i],
+        price: getRandomNum(500, 1000),
+        type: types[getRandomNum(0,4)],
+        rooms: getRandomNum(1, 7),
+        guests: getRandomNum(1, 7),
+        checkin: checkinArr[getRandomNum(0,2)],
+        checkout: checkinArr[getRandomNum(0,2)],
+        features: featuresArr.slice(0, getRandomNum(1,5)),
+        description: descriptions[getRandomNum(1,16)],
+        photos: photosArr.slice(0, getRandomNum(1, 3)),
+      },
+
+      location : {
+        lat: getRandomFloatingPoint(35.65000, 35.70000, 5),
+        lng: getRandomFloatingPoint(139.70000, 139.80000, 5)
+      },
+    };
+
+    advert.offer.address = `${advert.location.lat}, ${advert.location.lng}`;
+    arr.push(advert);
+  }
+  return arr;
+};
+
+createAdvertArchive(10);
+
